No need to hide this function with 'new'.
This is a little iffy, semantically, because we are shadowing a function
that is in scope (since the subclass is nested), but its not derived,
since the superclass defines it as private.

The other option here could be to define it as protected in the
superclass, though that wouldn't really change anything other than the
compiler's worldview. Or just rename the inner class' function.